Improve job tests
Add individual jobs to the test suite
