+++ test/jobs/5990
@@ -1,0 +1,1 @@
+{"job":{"assets":{"hdd":["publiccloud_12sp4_GCE_BYOS_Updates.qcow2","publiccloud_tools_0018.qcow2"]},"assigned_worker_id":25,"blocked_by_id":null,"children":{"Chained":[5991],"Directly chained":[],"Parallel":[]},"clone_id":null,"group_id":null,"has_parents":0,"id":5990,"name":"sle-12-SP4-GCE-BYOS-Updates-x86_64-Build20210419-1-publiccloud_download_testrepos@gce_n1_standard_2","parents":{"Chained":[],"Directly chained":[],"Parallel":[]},"parents_ok":1,"priority":50,"result":"softfailed","settings":{"ARCH":"x86_64","ASMM_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Adv-Systems-Management_12_x86_64\/","BACKEND":"qemu","BOOT_HDD_IMAGE":"1","BUILD":"20210419-1","CONTM_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Containers_12_x86_64\/","DESKTOP":"textmode","DISTRI":"sle","FLAVOR":"GCE-BYOS-Updates","HDDSIZEGB":"20","HDD_1":"publiccloud_tools_0018.qcow2","INSTALLONLY":"1","ISO_MAXSIZE":"4700372992","KEEP_GRUB_TIMEOUT":"1","MACHINE":"gce_n1_standard_2","NAME":"00005990-sle-12-SP4-GCE-BYOS-Updates-x86_64-Build20210419-1-publiccloud_download_testrepos@gce_n1_standard_2","OS_TEST_ISSUES":"19064,19149","OS_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-SERVER_12-SP4_x86_64\/","PC_TOOLS_IMAGE_BASE":"publiccloud_tools_0018.qcow2","PUBLIC_CLOUD":"1","PUBLIC_CLOUD_BUILD":"1.20","PUBLIC_CLOUD_BUILD_KIWI":"0.9.4","PUBLIC_CLOUD_IMAGE_ID":"sles-12-sp4-byos-v20210303","PUBLIC_CLOUD_IMAGE_LOCATION":"http:\/\/dist.suse.de\/ibs\/SUSE:\/SLE-12-SP4:\/GA:\/TEST\/images\/SLES12-SP4-GCE-BYOS.x86_64-0.9.4-Build1.20.tar.gz","PUBLIC_CLOUD_IMAGE_NAME":"sles-12-sp4-byos-v20210303","PUBLIC_CLOUD_IMAGE_PROJECT":"suse-byos-cloud","PUBLIC_CLOUD_IMAGE_STATE":"inactive","PUBLIC_CLOUD_IMAGE_TYPE":"uefi","PUBLIC_CLOUD_INSTANCE_TYPE":"n1-standard-2","PUBLIC_CLOUD_PROVIDER":"GCE","PUBLIC_CLOUD_RESOURCE_NAME":"openqa-suse-de","PUBLIC_CLOUD_STORAGE":"openqa-qam-storage","PUBLIC_CLOUD_UPLOAD_INSTANCE_HELPER":"gce_n1_standard_2","PUBLIC_CLOUD_VAULT_NAMESPACE":"qac","PUBLISH_HDD_1":"publiccloud_12sp4_GCE_BYOS_Updates.qcow2","QEMUVGA":"cirrus","QEMU_DISABLE_SNAPSHOTS":"1","REGISTRY":"18.156.2.117:5000","REPOHASH":"6b2b87d941fa970cd98f02b14444c27c","SCC_ADDONS":"sdk,tcm,wsm,contm,asmm,pcm,ltss","SCC_REGCODE":"deedc51104e549deb","SCC_REGCODE_LTSS":"INTERNAL-USE-ONLY-cb52-2b0b","SDK_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-SDK_12-SP4_x86_64\/","TCM_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Toolchain_12_x86_64\/","TEST":"publiccloud_download_testrepos","TEST_SUITE_NAME":"publiccloud_download_testrepos","VERSION":"12-SP4","VIDEOMODE":"text","VIRTIO_CONSOLE":"1","WE_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-WE_12-SP4_x86_64\/","WORKER_CLASS":"pc_qam_gce","WSM_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Web-Scripting_12_x86_64\/","YAML_SCHEDULE":"schedule\/publiccloud\/qem_publiccloud_download_testrepos.yml","ZYPPER_ORPHANED_CHECK_ONLY":"1","ZYPPER_WHITELISTED_ORPHANS":"cracklib-dict-small,gio-branding-upstream,klogd,ksh,yast2-slp-server,python3-humanfriendly"},"state":"done","t_finished":"2021-04-19T12:40:53","t_started":"2021-04-19T12:34:25","test":"publiccloud_download_testrepos"}}
