+++ test/jobs/5993
@@ -1,0 +1,1 @@
+{"job":{"assets":{"hdd":["publiccloud_12sp5_Azure_Basic_gen2_Updates.qcow2"]},"assigned_worker_id":25,"blocked_by_id":null,"children":{"Chained":[],"Directly chained":[],"Parallel":[]},"clone_id":null,"group_id":null,"has_parents":1,"id":5993,"name":"sle-12-SP5-AZURE-Basic-gen2-Updates-x86_64-Build20210419-1-publiccloud_boottime@az_Standard_B2s","origin_id":5992,"parents":{"Chained":[5985],"Directly chained":[],"Parallel":[]},"parents_ok":1,"priority":50,"result":"passed","settings":{"ARCH":"x86_64","AZURE_LOGIN_WAIT_SECONDS":"120","BACKEND":"qemu","BOOT_HDD_IMAGE":"1","BUILD":"20210419-1","BUILD_SDK":"GM","COLLECT_COREDUMPS":"1","DESKTOP":"textmode","DISTRI":"sle","FLAVOR":"AZURE-Basic-gen2-Updates","HDDSIZEGB":"20","HDD_1":"publiccloud_12sp5_Azure_Basic_gen2_Updates.qcow2","INSTALLONLY":"1","KEEP_GRUB_TIMEOUT":"1","LTP_KNOWN_ISSUES":"http:\/\/qam.suse.de\/media\/downloads\/ltp_known_issues.json","MACHINE":"az_Standard_B2s","MRU_ADDONS":"sdk,tcm,wsm,we,contm,asmm","NAME":"00005993-sle-12-SP5-AZURE-Basic-gen2-Updates-x86_64-Build20210419-1-publiccloud_boottime@az_Standard_B2s","OS_TEST_ISSUES":"19037,19062,19064,19075,19114,19122,19128,19143,19148","OS_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-SERVER_12-SP5_x86_64\/","PC_TOOLS_IMAGE_BASE":"publiccloud_tools_0018.qcow2","PUBLIC_CLOUD":"1","PUBLIC_CLOUD_AZURE_OFFER":"sles-12-sp5-basic","PUBLIC_CLOUD_AZURE_SKU":"gen2","PUBLIC_CLOUD_CHECK_BOOT_TIME":"1","PUBLIC_CLOUD_IMAGE_ID":"","PUBLIC_CLOUD_INSTANCE_TYPE":"Standard_B2s","PUBLIC_CLOUD_PERF_DB_URI":"http:\/\/openqa-perf.qa.suse.de:8086","PUBLIC_CLOUD_PROVIDER":"AZURE","PUBLIC_CLOUD_QAM":"1","PUBLIC_CLOUD_RESOURCE_NAME":"openqa-suse-de","PUBLIC_CLOUD_VAULT_NAMESPACE":"qac","QAM_PUBLICCLOUD_SKIP_DOWNLOAD":"1","QA_HEAD_REPO":"http:\/\/10.160.0.100\/ibs\/QA:\/Head\/SLE-12-SP5\/","QA_WEB_REPO":"http:\/\/10.160.0.100\/ibs\/QA:\/SLE12SP5\/update\/","QEMURAM":"2048","QEMUVGA":"cirrus","QEMU_DISABLE_SNAPSHOTS":"1","REGISTRY":"18.156.2.117:5000","REPOHASH":"8a7a7a1e551d84aabc006a054eb6b948","REQUIRED_ADDON":"","SCC_ADDONS":"sdk,tcm,wsm,contm,asmm,pcm","SCC_EMAIL":"osukup@suse.com","SCC_REGCODE":"INTERNAL-USE-ONLY-b6d61029abdfb29b","SCC_REGCODE_LIVE":"INTERNAL-USE-ONLY-bfebec99e2edcbf6","SCC_REGCODE_WE":"INTERNAL-USE-ONLY-3d6c4dd6a6bcca8c","SDK_TEST_ISSUES":"18953,19114,19128,19143,19148","SET_CUSTOM_PROMPT":"1","SHUTDOWN_NEEDS_AUTH":"1","START_AFTER_TEST":"publiccloud_download_testrepos","TEST":"publiccloud_boottime","TEST_SUITE_NAME":"publiccloud_boottime","VERSION":"12-SP5","VIDEOMODE":"text","VIRTIO_CONSOLE":"1","WORKER_CLASS":"pc_qam_azure"},"state":"done","t_finished":"2021-04-19T13:51:34","t_started":"2021-04-19T13:34:01","test":"publiccloud_boottime"}}
