+++ test/jobs/5995
@@ -1,0 +1,1 @@
+{"job":{"assets":{"hdd":["publiccloud_15sp2_EC2_ARM_Updates.qcow2"]},"assigned_worker_id":25,"blocked_by_id":null,"children":{"Chained":[],"Directly chained":[],"Parallel":[]},"clone_id":null,"group_id":null,"has_parents":1,"id":5995,"name":"sle-15-SP2-EC2-ARM-Updates-x86_64-Build20210419-1-publiccloud_ltp_cve@ec2_a1.large","parents":{"Chained":[5994],"Directly chained":[],"Parallel":[]},"parents_ok":1,"priority":50,"result":"passed","settings":{"ARCH":"x86_64","BACKEND":"qemu","BASE_TEST_ISSUES":"18906,18947,18948,18985,19005,19036,19046,19055,19060,19072,19095,19106,19107,19111,19112,19115,19117,19119,19121,19129,19136,19147","BASE_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Basesystem_15-SP2_x86_64\/","BOOT_HDD_IMAGE":"1","BUILD":"20210419-1","COMMAND_EXCLUDE":"cve-2015-3290|cve-2017-17805","COMMAND_FILE":"cve","DESKTOP":"textmode","DESKTOP_TEST_ISSUES":"18906,19079,19146","DESKTOP_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Desktop-Applications_15-SP2_aarch64\/","DISTRI":"sle","FLAVOR":"EC2-ARM-Updates","HDDSIZEGB":"20","HDD_1":"publiccloud_15sp2_EC2_ARM_Updates.qcow2","INSTALLONLY":"1","KEEP_GRUB_TIMEOUT":"1","LEGACY_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Legacy_15-SP2_aarch64\/","LTP_TIMEOUT":"5400","MACHINE":"ec2_a1.large","NAME":"00005995-sle-15-SP2-EC2-ARM-Updates-x86_64-Build20210419-1-publiccloud_ltp_cve@ec2_a1.large","OS_TEST_ISSUES":"","PCM_TEST_ISSUES":"19076,19109","PCM_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Public-Cloud_15-SP2_x86_64\/","PC_TOOLS_IMAGE_BASE":"publiccloud_tools_0018.qcow2","PUBLIC_CLOUD":"1","PUBLIC_CLOUD_ARCH":"arm64","PUBLIC_CLOUD_BUILD":"1.6","PUBLIC_CLOUD_BUILD_KIWI":"0.9.16","PUBLIC_CLOUD_IMAGE_ID":"ami-0b01dd8e9169233b8","PUBLIC_CLOUD_IMAGE_NAME":"suse-sles-15-sp2-v20210303-hvm-ssd-arm64","PUBLIC_CLOUD_IMAGE_STATE":"active","PUBLIC_CLOUD_INSTANCE_TYPE":"a1.large","PUBLIC_CLOUD_LTP":"1","PUBLIC_CLOUD_PROVIDER":"EC2","PUBLIC_CLOUD_QAM":"1","PUBLIC_CLOUD_REGION":"eu-central-1","PUBLIC_CLOUD_RESOURCE_NAME":"openqa-suse-de","PUBLIC_CLOUD_UPLOAD_INSTANCE_HELPER":"ec2_t2.large","PUBLIC_CLOUD_VAULT_NAMESPACE":"qac","QEMUVGA":"cirrus","QEMU_DISABLE_SNAPSHOTS":"1","REGISTRY":"18.156.2.117:5000","REPOHASH":"0b68d64be0ba0a622ff7058465d4cfef","SCC_REGCODE":"deedc51104e549deb","SDK_TEST_ISSUES":"19095,19121,19146","SDK_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Development-Tools_15-SP2_aarch64\/","SERVERAPP_TEST_ISSUES":"19015,19106,19129,19136","SERVERAPP_TEST_TEMPLATE":"http:\/\/download.suse.de\/ibs\/SUSE:\/Maintenance:\/@INCIDENTNR@\/SUSE_Updates_SLE-Module-Server-Applications_15-SP2_x86_64\/","START_AFTER_TEST":"publiccloud_download_testrepos","TEST":"publiccloud_ltp_cve","TEST_SUITE_NAME":"publiccloud_ltp_cve","VERSION":"15-SP2","VIDEOMODE":"text","VIRTIO_CONSOLE":"1","WORKER_CLASS":"pc_qam_ec2","ZYPPER_ORPHANED_CHECK_ONLY":"1","ZYPPER_WHITELISTED_ORPHANS":"opie,SuSEfirewall2"},"state":"done","t_finished":"2021-04-19T15:06:35","t_started":"2021-04-19T13:59:39","test":"publiccloud_ltp_cve"}}
