+++ test/jobs/overview
@@ -1,1 +1,1 @@
-[{"id":5136,"name":"sle-15-SP1-AZURE-BYOS-gen2-Updates-x86_64-BuildFalse-qem_publiccloud_img_proof@az_Standard_B2s"},{"id":5135,"name":"sle-15-SP1-AZURE-BYOS-gen2-Updates-x86_64-BuildFalse-mau-extratests-publiccloud@az_Standard_B2s"},{"id":5134,"name":"sle-15-SP1-AZURE-BYOS-gen2-Updates-x86_64-BuildFalse-mau-extratests-docker-publiccloud@az_Standard_B2s"},{"id":5133,"name":"sle-15-SP1-AZURE-BYOS-Updates-x86_64-BuildFalse-qem_publiccloud_img_proof@az_Standard_A2_v2"},{"id":5132,"name":"sle-15-SP1-AZURE-BYOS-Updates-x86_64-BuildFalse-mau-extratests-publiccloud@az_Standard_A2_v2"},{"id":5131,"name":"sle-15-SP1-AZURE-BYOS-Updates-x86_64-BuildFalse-mau-extratests-docker-publiccloud@az_Standard_A2_v2"}]
+[{"id":5995,"name":"sle-15-SP2-EC2-ARM-Updates-x86_64-Build20210419-1-publiccloud_ltp_cve@ec2_a1.large"},{"id":5994,"name":"sle-15-SP2-EC2-ARM-Updates-x86_64-Build20210419-1-publiccloud_download_testrepos@ec2_a1.large"},{"id":5993,"name":"sle-12-SP5-AZURE-Basic-gen2-Updates-x86_64-Build20210419-1-publiccloud_boottime@az_Standard_B2s"},{"id":5991,"name":"sle-12-SP4-GCE-BYOS-Updates-x86_64-Build20210419-1-publiccloud_ltp_cve@gce_n1_standard_2"},{"id":5990,"name":"sle-12-SP4-GCE-BYOS-Updates-x86_64-Build20210419-1-publiccloud_download_testrepos@gce_n1_standard_2"},{"id":5985,"name":"sle-12-SP5-AZURE-Basic-gen2-Updates-x86_64-Build20210419-1-publiccloud_download_testrepos@az_Standard_B2s"}]
